requirements.txt
anthropic>=0.40.0
beautifulsoup4>=4.12.0
praw>=7.7.0
requests>=2.31.0
pyyaml>=6.0
lxml>=4.9.0
python-dateutil>=2.8.0
